Refactor: 사용하지 않는 domain 모듈 종속성 주석처리, 비즈니스 로직 클래스 도메인 모듈로 이동
