+++ closet-domain/build.gradle
@@ -10,7 +10,7 @@ project(':closet-domain') {
         implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
         testImplementation 'com.h2database:h2'
         runtimeOnly 'org.postgresql:postgresql'
-        implementation 'org.springframework.boot:spring-boot-starter-data-redis'
+//        implementation 'org.springframework.boot:spring-boot-starter-data-redis'
 
     }
 }
