+++ closet-domain/src/main/java/project/closet/weather/WeatherData.java
@@ -1,0 +1,16 @@
+package project.closet.weather;
+
+import lombok.Builder;
+
+@Builder
+public record WeatherData(
+    String baseDate,   // 예보 발표 날짜
+    String baseTime,   // 예보 발표 시각
+    String category,   // 항목 코드 (TMP, SKY, etc)
+    String fcstDate,   // 예보 날짜
+    String fcstTime,   // 예보 시각
+    String fcstValue,  // 값
+    int nx,
+    int ny
+) {
+}
